requirements.txt
python-telegram-bot[job-queue]>=21.0
requests
google-generativeai
telegram
